move components back to original folder
